Override BlackDuck Project Name and Project Version in Image
The project name and version used in BlackDuck could be customized
previously with caveats:
* The BlackDuck Project Name had to be the same as the Repository
* The BlackDuck Project Version was constructed using the Tag + "-" +
first 20 characters of the Sha

This change allows optional BlackDuck Project Name and BlackDuck Project
Version to be sent in the Image data block of the request body. The Name
and Version provided will override the default functionality previously
described.